Tweak

+++ src/include/access/relscan.h
@@ -131,7 +131,7 @@ typedef struct IndexScanDescData
 										 * tuples */
 
 	/* index access method instrumentation output state */
-	double		nprimscans;		/* # of primitive index scans */
+	uint64 		nprimscans;		/* # of primitive index scans */
 
 	/* index access method's private state */
 	void	   *opaque;			/* access-method-specific info */
